Fix Commentary entity and repository

+++ src/main/java/com/alkemy/ong/data/repository/CommentaryRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentaryRepository extends JpaRepository<Commentary, Integer> {
+public interface CommentaryRepository extends JpaRepository<Commentary, Long> {
+
 }
